wrap up resource views feature
