+++ spec/support/pages/grove_admin_page.rb
@@ -1,5 +1,30 @@
+require_relative '../pages/page'
 module Pages
   class GroveAdminPage < Page
+
+    def visit_page
+      visit '/admin/groves'
+      self
+    end
+
+    def view_grove(name)
+      within("tr", text: name) do
+        click_on "View"
+      end
+      self
+    end
+
+    def go_to_show_page(id)
+      visit "/admin/groves/#{id}"
+      self
+    end
+
+    def delete_grove(id)
+      within("#grove-#{id}") do
+        click_on('delete')
+      end
+    end
+
     def create_new_grove(name="Ponderosa")
       page.has_content? "Create A New Grove"
       page.fill_in "Name", with: name
